Extract job factory to separate class
